split blockedapps into other package
